@@ -21,33 +21,35 @@
   },
   "dependencies": {
     "@angular/animations": "^9.1.11",
-    "@angular/common": "^8.1.0",
-    "@angular/compiler": "^8.1.0",
-    "@angular/core": "^8.1.0",
-    "@angular/forms": "^8.1.0",
-    "@angular/platform-browser": "^8.1.0",
-    "@angular/platform-browser-dynamic": "^8.1.0",
-    "@angular/router": "^8.1.0",
+    "@angular/common": "^9.1.11",
+    "@angular/compiler": "^9.1.11",
+    "@angular/core": "^9.1.11",
+    "@angular/forms": "^9.1.11",
+    "@angular/localize": "^9.1.11",
+    "@angular/platform-browser": "^9.1.11",
+    "@angular/platform-browser-dynamic": "^9.1.11",
+    "@angular/router": "^9.1.11",
     "@fortawesome/fontawesome-free": "^5.1.0",
     "@ng-bootstrap/ng-bootstrap": "^5.0.0-rc.1",
     "@ngx-translate/core": "^11.0.1",
     "bootstrap": "^4.1.1",
     "ngx-toastr": "^12.1.0",
     "rxjs": "^6.5.2",
-    "zone.js": "^0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.801.0",
-    "@angular/cli": "~8.1.0",
-    "@angular/compiler-cli": "^8.1.0",
-    "@angular/language-service": "^8.1.0",
+    "@angular-devkit/build-angular": "~0.901.8",
+    "@angular/cli": "~9.1.8",
+    "@angular/compiler-cli": "^9.1.11",
+    "@angular/language-service": "^9.1.11",
     "@angularclass/hmr": "^2.1.3",
     "@biesbjerg/ngx-translate-extract": "^2.3.4",
     "@ngx-rocket/scripts": "^4.0.0",
     "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^10.9.0",
-    "codelyzer": "^5.1.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "hads": "^1.7.0",
     "htmlhint": "^0.11.0",
     "https-proxy-agent": "^2.0.0",
@@ -73,7 +75,7 @@
     "ts-node": "^8.3.0",
     "tslint": "~5.18.0",
     "tslint-config-prettier": "^1.14.0",
-    "typescript": "~3.4.0"
+    "typescript": "~3.8.3"
   },
   "prettier": {
     "singleQuote": true,
