@@ -7,7 +7,7 @@ import { RamdomService } from './ramdom.service';
   imports: [CommonModule]
 })
 export class ServicesModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<ServicesModule> {
     return {
       ngModule: ServicesModule,
       providers: [EntriesService, RamdomService, EntriesService]
